+++ incapsula/operation_constants.go
@@ -25,6 +25,11 @@ const CreateSubAccount = "create_sub_account"
 const ReadSubAccount = "read_sub_account"
 const DeleteSubAccount = "delete_sub_account"
 
+const CreateWAFLogSetup = "create_waf_log_setup"
+const DeleteWAFLogSetup = "delete_waf_log_setup"
+const ActivateWAFLogSetup = "activate_waf_log_setup"
+const UpdateStatusWAFLogSetup = "update_status_waf_log_setup"
+
 const ReadAccountDataStorageRegion = "read_account_data_storage_region"
 const UpdateAccountDataStorageRegion = "update_account_data_storage_region"